Rename modules of photography
