+++ src/components/photography/more.js
